Move all Nunjucks config to its own function
This adds all the Nunjucks config to its own function,
as we'll eventually want to add a bunch more global
helpers to our views, so this keeps our main.ts file
a bit tidier.
